some refactoring; moving to token balance schema
