+++ subgraph.yaml
@@ -29,3 +29,24 @@ dataSources:
           handler: handleCreatedEdition
           receipt: true
       file: ./src/event-handlers.ts
+
+templates:
+  - kind: ethereum/contract
+    name: SingleEditionMintable
+    network: matic
+    source:
+      abi: SingleEditionMintable
+    mapping:
+      kind: ethereum/events
+      apiVersion: 0.0.7
+      language: wasm/assemblyscript
+      entities:
+        - FreeNFTTransfer
+      abis:
+        - name: SingleEditionMintable
+          file: ./abis/SingleEditionMintable.json
+      eventHandlers:
+        - event: Transfer(indexed address,indexed address,indexed uint256)
+          handler: handleTransfer
+          receipt: true
+      file: ./src/event-handlers.ts
