+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
   "extends": "@graphprotocol/graph-ts/types/tsconfig.base.json",
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": ["node_modules"]
 }
